Refactor union-find classes to introduce AbstractSlottedUnionFind and streamline SlottedUnionFind implementation
