+++ foresight/src/main/scala/foresight/eqsat/hashCons/AbstractMutableSlottedUnionFind.scala
@@ -7,7 +7,7 @@ import foresight.eqsat.collections.{SlotMap, SlotSet}
  * A mutable union-find data structure that supports path compression and union by rank. The keys are e-class references
  * and the values are e-class calls, which include the renaming of the slots.
  */
-private[hashCons] abstract class AbstractMutableSlottedUnionFind {
+private[hashCons] abstract class AbstractMutableSlottedUnionFind extends AbstractSlottedUnionFind {
   /**
    * Updates the parent of the given key to the given value.
    *
@@ -15,25 +15,6 @@ private[hashCons] abstract class AbstractMutableSlottedUnionFind {
    * @param value The new parent of the key.
    */
   def update(key: EClassRef, value: EClassCall): Unit
-
-  /**
-   * Finds the representative of the given key. If the key is not in the union-find, null is returned; otherwise, the
-   * representative of the key is returned.
-   *
-   * @param ref The key to find.
-   * @return The representative of the key, if the key is in the union-find. Null otherwise.
-   */
-  protected def getParentOrNull(ref: EClassRef): EClassCall
-
-  /**
-   * Checks if the given e-class is the canonical representative of its set. Assumes that the e-class is in the union-find.
-   * @param ref The e-class to check.
-   * @return True if the e-class is the canonical representative of its set, false otherwise.
-   */
-  final def isCanonical(ref: EClassRef): Boolean = {
-    // require(parents.contains(ref), s"EClassRef $ref must be in the union-find.")
-    getParentOrNull(ref).ref == ref
-  }
 
   /**
    * Adds a new key to the union-find with itself as its own parent and the given slots.
